backend/requirements.txt
pydantic-settings~=2.11.0
SQLAlchemy~=2.0.43
celery~=5.5.3
uvicorn~=0.37.0
fastapi~=0.118.0
pydantic~=2.11.10
psycopg2-binary~=2.9.10
redis~=6.4.0
google-genai~=1.41.0